DevOps: Fix the CI workflow and run the unit tests
The GHA workflow is renamed to `ci.yml` and is made into a proper
continuous integration workflow. The old one only ran the pre-commit for
a matrix of Python version, but this is superfluous.

The workflow is split in seperate jobs. The `pre-commit` runs the
pre-commit on a single Python version. The `tests` job runs the unit
test suite through `pytest` on a matrix of all supported Python
versions.