Do not retrieve gbw file by default
ORCA stores molecular wavefunction in a binary GBW file,
which can then be used to restart a calculation or used
as a wavefunction guess. Currently, in OrcaCalculation
we always fetch this file as part of the retrieved folder.
This is not ideal since these files can get very big and
it is not possible to delete them once the workflow finishes
since they become part of the AiiDA provenance.

I simply remove the GBW file from the default retrieve list.
User can still add it when needed via inputs.metadata.options.additional_retrieve_list