Remove inputs.settings Dict
The only remaining supported keyword in settings Dict was 'cmdline'
for passing extra command line parameters.
However, I had a look at ORCA manual and orca binary actually does not support
any other params besides the input file. The only exception are
extra MPI params, that must be passed **after** the input file,
not before, as was done here, and needs to be escaped by quotes.
See section "Hints of the use of Parallel ORCA" in the manual.

If we ever need to support this,
it's better to give it a different and more specific name anyway.
Until then let's remove the settings input to keep the code clean.